Update curseforge to reference a working modpack
Update wording to reference a modpack that works out the box as opposed to sevtech ages which doesn't work. I added my BOFA mods modpack however we can use any modpack that works out the box. If used, I vow to not break this modpack and will leave it in it's current state to guarantee it will work when people use it as a test.