Delete steamcmd_servers/arma/arma3_headless_client directory
This Egg may boot properly, but it appears that the Linux version of Arma's Headless Client will not actually connect properly to the main server unless it is on the local host of the main server (connecting via 127.0.0.1). Therefore, this issue, along with HC functionality being added to the main Arma 3 Egg, I propose that this egg be retired.