Fix tshock egg stuck on reinstallation
The tshock egg currently uses `unzip` command in the installation script to extract the downloaded files.

If the user reinstalls the server, the extracted files already exist, and `unzip` prompts for whether to overwrite the files or not. This causes the installation script to get stuck.

This fixes the problem by adding the `-o` option to overwrite the files without prompting.