Update ports to display the correct connection port
Previously the port that is displayed to clients was the telemetry port causing people to be confused as this isn't a port you can connect to.

This update changes the primary port to the server query port instead so the correct connection port is displayed, and it moves the telemetry port to the back end.